Allow passing URLSearchParams when searching for data
While the strongly-typed interface is useful in most cases, it does not support
some things, such as passing "id" multiple times in order to perform an "OR"
query.

Adding this to the "findData" API now since bulk fetching of data is a very
common case - but we may want to consider adding it to all/most getters.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>